Added simplified/optimized version of Core
This is the simplified version of PBKDF2 on JavaCard. It does not support any other hash functions than Sha1 and Sha256. It does not support outputs longer than hash outputs (sha1 - 20 bytes, sha256 - 32 bytes). But obviously it can be used several times and the result then concatenated, which is exactly what extended pbkdf2 does.
This is pretty much optimized for speed, but still it is a very lengthy process, because it requires a lot of iterations
on a slow CPU with very limited memory.

Usage: Either use original PBKDF2Core.java, or remove the original and rename this to PBKDF2Core.java.
There should be only 2 files in the src/ folder - the Core and an example Applet.

If you know of any speed/memory optimizations, please let me know, I'll surely acknowledge you in my diploma thesis.