MORUS: do not needlessly allocate heap memory
All arrays are, by default, allocated on the heap.  Therefore, for every
call of the morus_initialization method, a new set of constant arrays
was allocated making the repeated initialisation not feasible, especially
on cards with unreliable garbage collection.

Make these constant arrays class members instead so that they are allocated
only once.

Detected-by: JCProfilerNext